Move register_bot_tools to automated_recon directory and update import path
