+++ mcp_tools/automated_recon/bot.py
